requirements.txt
pandas==2.3.3
numpy==2.0.2
matplotlib==3.10.0
torch==2.5.1
torchvision==0.20.1
torchmetrics==1.8.2
opencv-python-headless==4.10.0.84
Pillow==11.3.0
tqdm==4.67.1
streamlit==1.54.0